Rename GenBank.py to be consistent with rest of package
